requirements.txt
pip==21.0.1
setuptools==54.0.0
wheel==0.36.2
Flask==1.0
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
PyYAML==5.3